% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_littersize.R
\name{Rep_littersize}
\alias{Rep_littersize}
\title{Litter Size Estimation}
\usage{
Rep_littersize(
  ReproData,
  NDay = 7,
  ParentPercDam = 80,
  ParentPercSire = 80,
  MinNLitter = 30
)
}
\arguments{
\item{ReproData}{\code{data.frame}. A dataset including at least the following columns:
\emph{AnimalAnonID}, \emph{ParentAnonID}, \emph{ParentType}, \emph{Probability},
\emph{Offspring_BirthDate}, \emph{Offspring_Inst}, and \emph{ParentAge}.}

\item{NDay}{\code{numeric}. Maximum number of consecutive days within which offspring are considered to belong to the same litter or clutch (default is 7).}

\item{ParentPercDam}{\code{numeric}. Minimum parentage probability (in \%) required to include a Dam (default is 80).}

\item{ParentPercSire}{\code{numeric}. Minimum parentage probability (in \%) required to include a Sire (default is 80).}

\item{MinNLitter}{\code{numeric}. Minimum number of litters required to compute summary statistics. The detailed litter data frame will be returned in all cases (default is 30).}
}
\value{
A list with the following components:
\item{summary}{A named list with litter size statistics:
\describe{
\item{NOffsp_prob}{Number of offspring with qualifying parentage probability}
\item{NParent_prob}{Number of dams (mothers) with qualifying parentage probability}
\item{NReprEvent}{Total number of reproductive events (litters)}
\item{MeanLittSize}{Mean litter size}
\item{MedLittSize}{Median litter size}
\item{SELittSize}{Standard error of litter size}
\item{analyzed}{Logical. Whether summary statistics could be calculated (TRUE) or not (FALSE)}
\item{error}{Character string explaining why summary stats were not calculated, if applicable}
\item{Nerr}{Error code: 0 = no error, 4 = no births meet criteria, 5 = not enough litters}
}}
\item{littSizeDf}{A data frame with one row per litter, including litter size and associated parental information.}
\item{littSizeTab}{A table of litter sizes: frequency and proportion for each size.}
\item{littSizeperAge}{A data frame showing mean and standard deviation of litter size by maternal age.}
}
\description{
Creates a data frame of litter size per mother and birth date, and generates summary tables of litter size statistics.
Only offspring-mother associations with sufficient parentage probability are included.
}
\examples{
data(toy_core)
data(toy_collection)
data(toy_parent)


#prepare Data
Data <- Rep_prepdata (coresubset = toy_core, offspring = toy_core, toy_collection, toy_parent, 
                      MinNRepro = 1, MinNPaRepro = 1,
                      BirthType_parent = "All",
                      BirthType_offspring = "All" )

#Estimate litter size
out <- Rep_littersize(Data$ReproData, MinNLitter = 5,
                      NDay = 7)
print(out$summary)
}
