% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_main.R
\name{Rep_main}
\alias{Rep_main}
\title{Reproductive Summary Master Function}
\usage{
Rep_main(
  coresubset,
  offspring,
  collection,
  parent,
  contraception = NULL,
  RepSect = c("agemat", "litter", "season", "weight1st", "pattern", "LRS", "IBI", "fert"),
  BirthType_parent = "Captive",
  BirthType_offspring = "Captive",
  Global = TRUE,
  MinInstitution = 2,
  MinNRepro = 100,
  MinNPaRepro = 30,
  ParentPerc = 80,
  MinNLitter = 20,
  NDay = 7,
  institution = NULL,
  lx = NULL,
  GrowthData = NULL,
  core = NULL,
  MinDate = "1980-01-01",
  NMinEvenperAge = 3,
  MinNImm = 0.2,
  UncertBirth = 365,
  maxweight = NULL,
  Taxa = NULL,
  MinNSeas = 50,
  RemContra = FALSE
)
}
\arguments{
\item{coresubset}{A \code{data.frame} of individual-level records including parents. Must include columns: \code{AnimalAnonID}, \code{BirthType}, \code{MaxBirthDate}, \code{MinBirthDate}, \code{DepartDate}, \code{BirthDate}, \code{SexType}, \code{FirstHoldingInstitution}, and \code{GlobalStatus}.}

\item{offspring}{A \code{data.frame} of individual-level records including offspring. Must include columns: \code{AnimalAnonID}, \code{BirthType}, \code{MaxBirthDate}, \code{MinBirthDate}, \code{DepartDate}, \code{BirthDate}, \code{SexType}, \code{FirstHoldingInstitution}, and \code{GlobalStatus}.}

\item{collection}{A \code{data.frame} with at least: \code{AnimalAnonID}, \code{ScopeType}, \code{ChangeDate}.}

\item{parent}{A \code{data.frame} of parentage information including: \code{ParentAnonID}, \code{AnimalAnonID}, \code{Probability}, \code{ParentCollectionScopeType}, \code{OffspringCollectionScopeType}, and \code{ParentOriginType}.}

\item{contraception}{\code{data.frame}. Contraception records including columns:
\emph{AnimalAnonID}, \emph{Date}, \emph{Status}. (default: \code{NULL}).}

\item{RepSect}{A character vector of sections to run. Options: \code{"agemat"}, \code{"litter"}, \code{"IBI"}, \code{"fert"}, \code{"season"}, \code{"pattern"}, \code{"LRS"}, \code{"weightrepro"}, \code{"weight1st"}.}

\item{BirthType_parent}{Filter for parents by birth type ("Captive", "Wild", "All"). Default = "Captive".}

\item{BirthType_offspring}{Filter for offspring by birth type. Default = "Captive".}

\item{Global}{Logical; if \code{TRUE}, only use globally scoped records. Default = TRUE.}

\item{MinInstitution}{Minimum number of institutions required for analysis to proceed. Default = 2.}

\item{MinNRepro}{Minimum number of reproductive events required. Default = 100.}

\item{MinNPaRepro}{Minimum number of reproductive parents required. Default = 30.}

\item{ParentPerc}{Minimum parentage probability (\%). Default = 80.}

\item{MinNLitter}{Minimum number of litters for \code{Rep_littersize}. Default = 20.}

\item{NDay}{Maximum days between births to consider same litter. Default = 7.}

\item{institution}{\code{data.frame} of institution AnonID and hemisphere to evaluate seasonality for seasonality analysis.}

\item{lx}{A named numeric vector of survivorship (used in \code{Rep_LRS}).}

\item{GrowthData}{A \code{data.frame} with individual growth measurements including: \code{AnimalAnonID}, \code{MeasurementDate}, \code{MeasurementValue}, \code{MeasurementType}, \code{MeasurementDateEstimateEnd}, \code{MeasurementDateEstimateStart}, and \code{RecordType}. (used in \code{Rep_weightrepro} and \code{Rep_weight1strepro}).}

\item{core}{A \code{data.frame} with individual-level metadata including:
\code{AnimalAnonID}, \code{BirthDate}, \code{BirthObserved}, and \code{BirthDateEstimateType}. The birth date is not necessarily known (used in \code{Rep_weight1strepro}).}

\item{MinDate}{\code{character 'YYYY-MM-DD'} Earliest offspring birth date to include data}

\item{NMinEvenperAge}{Minimum number of reproductive events per age class (used in \code{Rep_pattern}). Default = 3.}

\item{MinNImm}{Numeric. Minimum proportion of immature and mature individuals required (default = 0.2, used in \code{Rep_weight1strepro}).}

\item{UncertBirth}{Maximum uncertainty tolerated for dates (used in \code{Rep_weightrepro} and \code{Rep_weight1strepro}).}

\item{maxweight}{Optional. Numeric. Maximum plausible body mass to retain (defaults to taxon-specific value, used for \code{Rep_weightrepro} and \code{Rep_weight1strepro}).}

\item{Taxa}{Character. Taxonomic group. One of: \code{"Mammalia"}, \code{"Aves"}, \code{"Reptilia"},
\code{"Amphibia"}, \code{"Chondrichthyes"}, or \code{"Osteichthyes"}.}

\item{MinNSeas}{Minimum number of births for seasonality analysis. Default = 50.}

\item{RemContra}{\code{logical}. Whether contraception data should be used to remove individuals under contraception (see details) (default: \code{FALSE}).}
}
\value{
A named list with:
\describe{
\item{\code{summary}}{General metadata and section-by-section success flags}
\item{\code{agemat}}{(optional) Output from \code{Rep_agemat}}
\item{\code{litter}}{(optional) Output from \code{Rep_littersize}}
\item{\code{IBI}}{(optional) Output from \code{Rep_IBI}}
\item{\code{fert}}{(optional) Output from \code{Fep_fert}}
\item{\code{season}}{(optional) Output from \code{Rep_season}}
\item{\code{pattern}}{(optional) Output from \code{Rep_pattern}}
\item{\code{LRS}}{(optional) Output from \code{Rep_LRS}}
\item{ \code{weightrepro}}{(optional) Output from \code{Rep_weightrepro}}
\item{\code{weight1st}}{(optional) Output from \code{Rep_weight1strepro}
}}
}
\description{
Runs all selected reproductive analyses on a dataset of animal records, parentage, movement, collection, and optionally survival, growth, and institution information.
}
\details{
Contraception: An individual is removed for the rolling year around the dates when it is recorded under Active Contraception.
}
\examples{
data(toy_core)
data(toy_collection)
data(toy_parent)
data(toy_weights)
lx = seq(1,0,length.out = 50)
names(lx)= c(1:50)

institution = data.frame(
  AnonInstitutionID = unique(toy_core$FirstHoldingInstitution), 
  Latitude = 57)
out <- Rep_main (coresubset = toy_core, offspring = toy_core, collection = toy_collection, 
                 parent = toy_parent, institution = institution,
                 GrowthData = toy_weights,core = toy_core,lx = lx,
                 MinNRepro = 1, MinNPaRepro = 3,
                 BirthType_parent = "Captive", BirthType_offspring = "Captive",
                Taxa = "Reptilia"
) 
}
