% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_season.R
\name{Rep_season}
\alias{Rep_season}
\title{Estimate Birth Seasonality by Hemisphere}
\usage{
Rep_season(Data, Institution, MinNBirth = 10, Hemisphere = "All")
}
\arguments{
\item{Data}{A \code{data.frame} of reproductive records with columns:
\describe{
\item{\code{Offspring_BirthDate}}{Date of birth (class \code{Date}).}
\item{\code{Offspring_Inst}}{Institution ID where the birth occurred.}
}}

\item{Institution}{A \code{data.frame} with:
\describe{
\item{\code{AnonInstitutionID}}{Institution anonymous ID.}
\item{\code{Latitude}}{Latitude of the institution (used to assign hemisphere).}
}}

\item{MinNBirth}{\code{numeric}. Minimum number of births required to run analysis for each group (default = 10).}

\item{Hemisphere}{\code{character}. Which hemisphere(s) to analyze. One or more of:
\code{"North"}, \code{"South"}, \code{"All"} (default = "All").}
}
\value{
A named \code{list} with:
\describe{
\item{\code{summary}}{A list of statistics by hemisphere:
\itemize{
\item{\code{Nbirth}}: Number of births considered
\item{\code{MeanDayOfYear}}: Mean day of the year (linear)
\item{\code{MedianDayOfYear}}: Median birth day
\item{\code{L80}}: Range of dates containing 80\% of births (linear)
\item{\code{Range}}: Full range of birth days (linear)
\item{\code{MeanCircularDayOfYear}}: Mean date on the circular scale (0–365)
\item{\code{CircularRange80}}: Circular 80\% quantile range (converted to days)
\item{\code{RangeCircular}}: Full range of circular birth dates (in days)
\item{\code{analyzed}}: Logical indicating whether results are valid
\item{\code{error}}: Message in case of insufficient data
}}
\item{\code{monthly_table}}{A table of monthly birth counts and proportions for each hemisphere}
}
}
\description{
Computes descriptive statistics on birth date seasonality across one or both hemispheres,
including mean, median, monthly distribution, and concentration of birth dates.It uses julian dates but also
Uses circular statistics to account for the periodic nature of time (day of year). The function aligns hemispheres based on institutional latitude.
}
\examples{

# Example with simulated data
data_births <- data.frame(
  Offspring_BirthDate = as.Date("2020-01-01") + sample(0:364, 50, replace = TRUE),
  Offspring_Inst = sample(LETTERS[1:3], 50, replace = TRUE)
)
institutions <- data.frame(
  AnonInstitutionID = LETTERS[1:3],
  Latitude = c(45, -33, 10)
)

res <- Rep_season(data_births, institutions, MinNBirth = 10, Hemisphere = c("North", "South"))
# View seasonal summary
res$summary$North
}
