% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_age.R
\name{Sur_age}
\alias{Sur_age}
\title{Age-specific survival}
\usage{
Sur_age(Lx, xv, Nyear = 1)
}
\arguments{
\item{Lx}{\code{numeric} Matrix of posterior survivorship trajectories}

\item{xv}{\code{numeric}  Age vector}

\item{Nyear}{\code{numeric} number of years to survive Default = 1#}
}
\value{
a data frame including age, the mean and 95\% credible interval of the age specific probability to survive \code{Nyear} years.
}
\description{
Estimate age-specific survival, i.e. the probability of surviving at least Nyear additional years, from the parameters outputs of a Basta model.
}
\examples{
Lx = matrix(c(seq(1,0,by = -0.1), 1,seq(0.5,0,length.out = 10)),nrow =2)
out <- Sur_age(Lx, Nyear = 5, xv = c(0:10))
}
