% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_ple.R
\name{Sur_ple}
\alias{Sur_ple}
\title{Kaplan-Meier estimator}
\usage{
Sur_ple(Data)
}
\arguments{
\item{Data}{\code{data.frame} including at least the following columns \emph{deparAge}, \emph{entryAge} (\code{date}), and \emph{DepartType}}
}
\value{
A data frame with columns:
\itemize{
\item Ages: event times
\item ple: Kaplan-Meier survivorship estimate
\item event : number of deaths at each age
\item Snd: Greenwood-type cumulative variance term used downstream
}
}
\description{
Compute the Kaplan-Meier product-limit estimator from individual entry and departure ages.
}
\examples{

entryAge = sample(c(1:10), 200, replace = TRUE)
data <- data.frame(
  entryAge = entryAge,
  deparAge =  entryAge + sample(c(0:10), 200, replace = TRUE),
  DepartType = sample(c('C', 'D'), 200, replace = TRUE))


out<-Sur_ple(data)
}
