% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_xx.R
\name{Sur_xx}
\alias{Sur_xx}
\title{Survivorship quantile}
\usage{
Sur_xx(Lx, xv, xx = 0.5)
}
\arguments{
\item{Lx}{\code{numeric} Survivorship}

\item{xv}{\code{numeric} Age vector}

\item{xx}{\code{numeric} Proportion of the population still alive. Between 0 and 1.}
}
\value{
a list including the mean and 95\% credible interval age at which xx\% of the population is still alive
}
\description{
Estimate the age at which survivorship reaches a specified level xx
}
\examples{
Lx = matrix(c(seq(1,0,by = -0.1), 1,seq(0.5,0,length.out = 10)),nrow =2)
out <- Sur_xx(Lx,  xv = c(0:10), xx = 0.1)
}
