% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_collection.R
\docType{data}
\name{toy_collection}
\alias{toy_collection}
\title{Collection data}
\format{
A data.frame object with 1382 rows and 5 variables:
\describe{
\item{AnimalAnonID}{Anonimized ID of the Animal}
\item{RecordingInstitution}{Anonimized ID of the Institution}
\item{CollectionType}{Name of the collection}
\item{ScopeType}{"Local" = Only private use or "Global" = Zims user use}
\item{ChangeDate}{Date of arrival in this collection}
#' }
}
\source{
Subset of a toy database created by the Species360 Development team in 08-2024
}
\usage{
data(toy_collection)
}
\description{
A simulated dataset including information about individual collections.
}
\keyword{datasets}
