% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_deathinformation.R
\docType{data}
\name{toy_deathinformation}
\alias{toy_deathinformation}
\title{Death Information data}
\format{
A data.frame object with 266 rows and 3 variables:
\describe{
\item{AnimalAnonID}{Anonimized ID of the Animal}
\item{RelevantDeathInformationTypeID}{ID of the cause of death}
\item{RelevantDeathInformationType}{Cause of death}
#' }
}
\source{
Subset of a toy database created by the Species360 Development team in 08-2024
}
\usage{
data(toy_deathinformation)
}
\description{
A simulated dataset including information about individual death information.
}
\keyword{datasets}
