% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_moves.R
\docType{data}
\name{toy_moves}
\alias{toy_moves}
\title{Moves data}
\format{
A data.frame object with 3 variables:
\describe{
\item{AnimalAnonID}{Anonimized ID of the Animal}
\item{Date}{Date of the move/transaction/event}
\item{To}{Anonimized ID of institution of arrival}
}
}
\source{
Subset of a toy database created by the Species360 Development team in 08-2024
}
\usage{
data(toy_moves)
}
\description{
A simulated dataset including information about individual moves between institutions.
}
\keyword{datasets}
