% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zimsSRGa-package.R
\docType{package}
\name{zimsSRGa-package}
\alias{zimsSRGa}
\alias{zimsSRGa-package}
\title{zimsSRGa: Survival, Reproduction, and Growth Analytics}
\description{
A repeatable way to perform Survival, Reproduction, and Growth analytics from Zims Data.
}
\author{
\strong{Maintainer}: Floriane Plard \email{floriane.plard@Species360.org} [contractor]

Other contributors:
\itemize{
  \item Species 360 \email{Support@Species360.org} [copyright holder]
  \item Fernando Colchero [contributor]
}

}
\keyword{internal}
