% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_out.R
\name{Sur_out}
\alias{Sur_out}
\title{Create output and plots for the survival analysis}
\usage{
Sur_out(
  out,
  MinAge = 0,
  MinMLE = 0.1,
  MaxLE = 2,
  MinLx = 0.1,
  MaxAge = 120,
  ncpus = 2,
  PlotDir = NULL,
  PlotName = ""
)
}
\arguments{
\item{out}{\code{list} output of the survival analysis,  returns from \code{Sur_ana()}}

\item{MinAge}{\code{numeric} Age at which the survival analysis should start, in years.  see ?basta for more information.}

\item{MinMLE}{\code{numeric} Value used for checks. Minimum survivorship allowed at mean life expectancy. Between 0 and 1.}

\item{MaxLE}{\code{numeric} Value used for checks. Maximum remaining life expectancy at last observed age. In years.}

\item{MinLx}{\code{numeric} Value used for longevity threshold and for checks. between 0 and 1. Minimum reached survivorship from the raw Kaplan-Meier analysis. This number avoids running survival analysis if there are too few dead individuals in the data. Lower is better.}

\item{MaxAge}{\code{numeric} Maximum possible age in years. Only used for model predictions. This argument is not used to select data.}

\item{ncpus}{\code{numeric} Number of CPU cores to use.}

\item{PlotDir}{\code{character} Directory to save the plots. If \code{NULL}, no plot is saved}

\item{PlotName}{\code{character} Name used to save the plots.}
}
\value{
The output list of the survival analysis including, in addition:
\itemize{
\item Key survival metrics including Mean life expectancy (MLE & Ex), median life
expectancy (L50) and  age at which 90\% of the individual died (Longevity = L90)
estimated from raw data, the Kaplan-Meier estimator, and the survival model. Estimates
include ages from birth or from age at sexual maturity (if given). First year
survival, First month survival, Entropy (H and Epx = -log(H)), coefficient of
variation (CV) and Gini coefficient (G) are also estimated from the Kaplan-Meier
estimator and the survival model.
\item Checks of the fit for the selected survival model:
\itemize{
\item Gof_KM_coeff1: Percentage of age points where the KM estimator is outside of the
95\% CI of the survivorship estimated from the survival model.
\item Gof_KM_coeff2: Maximum sum of same sign residuals between KM estimator and
survival model
\item Gof_Martingal_KM: Value of the test that must be compared to a kolmogorov Smirnov table and test telling if the model fits
\item LxatMLE: Estimated survivorship at mean life expectancy
\item LEmaxOage: Life expectancy at max observed age
\item KMMinLx: Minimum survivorship reached by the Kaplan-Meier estimator
}
\item The remaining life expectancy per age (relex_from0)
\item The probability to live 5 years more (Sur5)
\item Age-specific survival  (Sur1)
\item Monthly survival  (Sur1m)
\item Plots if requested
\itemize{
\item One showing the convergence of chains, together with estimated survivorship and
survival and mortality rates.
\item The second showing remaining life-expectancy, and the age-specific probability
to live 1 year more.
}
}
}
\description{
estimates key metrics from the selected survival model and returns illustrating plots of the survival model.
}
\examples{
data(core)
data(deathinformation)
out <- Sur_ana(core,  DeathInformation = deathinformation, 
               Models = "GO", Shape = "simple",
               niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3)

out <- Sur_out(out, ncpus = 3)
}
